performs refactoring
